requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-SocketIO==5.3.4
Flask-Migrate==4.0.5

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.20
cryptography==41.0.3

# Redis
redis==5.0.0
flask-caching==2.0.2

# WebSocket
python-socketio==5.9.0
python-engineio==4.7.1
eventlet==0.33.3
dnspython<2.3.0  # Pin to avoid eventlet compatibility issues with ignore_errors parameter

# Security
Werkzeug==2.3.7
bcrypt==4.0.1

# Forms and validation
WTForms==3.0.1
email-validator==2.0.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# WSGI Server
gunicorn==21.2.0
